# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flash API",
    "version": "0.1.0",
    "description": "API de la plateforme de monnaie électronique Flash."
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "paths": {
...
    "/v1/merchant/charges": {
      "post": {
        "summary": "Créer un QR dynamique (montant + référence + expiration)",
        "tags": [
          "merchants"
        ],
        "responses": {
          "201": {
            "description": "Succès"
          },
          "4XX": {
            "description": "Erreur métier ou de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "amount_minor": {
                    "examples": [
                      25000
                    ],
                    "exclusiveMinimum": 0,
                    "title": "Amount Minor",
                    "type": "integer"
                  },
                  "reference": {
                    "examples": [
                      "Table 4"
                    ],
                    "maxLength": 80,
                    "minLength": 1,
                    "title": "Reference",
                    "type": "string"
                  },
                  "ttl_minutes": {
                    "default": 60,
                    "maximum": 1440,
                    "minimum": 1,
                    "title": "Ttl Minutes",
                    "type": "integer"
                  }
                },
                "required": [
                  "amount_minor",
                  "reference"
                ],
                "title": "CreateChargeRequest",
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/v1/merchant/payments": {
      "get": {
        "summary": "Encaissements du marchand",
        "tags": [
          "merchants"
        ],
        "responses": {
          "200": {
            "description": "Succès"
          },
          "4XX": {
            "description": "Erreur métier ou de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/merchant/v1/charges": {
      "post": {
        "summary": "API marchande : créer une demande de paiement (QR dynamique)",
        "tags": [
          "merchant-api"
        ],
        "responses": {
          "201": {
            "description": "Succès"
          },
          "4XX": {
            "description": "Erreur métier ou de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "amount_minor": {
                    "examples": [
                      25000
                    ],
                    "exclusiveMinimum": 0,
                    "title": "Amount Minor",
                    "type": "integer"
                  },
                  "reference": {
                    "examples": [
                      "Commande #4821"
                    ],
                    "maxLength": 80,
                    "minLength": 1,
                    "title": "Reference",
                    "type": "string"
                  },
                  "ttl_minutes": {
                    "default": 60,
                    "maximum": 1440,
                    "minimum": 1,
                    "title": "Ttl Minutes",
                    "type": "integer"
                  }
                },
                "required": [
                  "amount_minor",
                  "reference"
                ],
                "title": "CreateChargeBody",
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/merchant/v1/charges/{charge_id}": {
      "get": {
        "summary": "API marchande : état d'une demande de paiement",
        "tags": [
          "merchant-api"
        ],
        "responses": {
          "200": {
            "description": "Succès"
          },
          "4XX": {
            "description": "Erreur métier ou de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/merchant/v1/charges/{charge_id}/refund": {
      "post": {
        "summary": "API marchande : rembourser le paiement d'une demande",
        "tags": [
          "merchant-api"
        ],
        "responses": {
          "200": {
            "description": "Succès"
          },
          "4XX": {
            "description": "Erreur métier ou de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 255
            }
          }
        ]
      }
    },
    "/merchant/v1/payments": {
      "get": {
        "summary": "API marchande : encaissements récents",
        "tags": [
          "merchant-api"
        ],
        "responses": {
          "200": {
            "description": "Succès"
          },
          "4XX": {
            "description": "Erreur métier ou de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message",
          "details"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "INSUFFICIENT_FUNDS"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}